Only allow one publicHearingNotice, not an array.
Notices are used elsewhere. If we mix array and object references, then
versioning becomes tricky and we lose a lot of granularity in the versioning.
As we don't have much planning data, and we know we want to see the notice
changes in as much detail as possible, am making publicHearingNotices singular
until we need otherwise. In the event that we do need more than one
publicHearingNotice, we may well just want to create something different and
not re-use notice.